requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
celery==5.3.6
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-docx==1.1.0
pdfplumber==0.10.0
pdfminer.six==20231228
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
langdetect==1.0.9
cryptography==42.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
structlog==24.1.0
